+++ spa-challenge/src/views/Search/style.scss
@@ -1,0 +1,3 @@
+@import '@/styles/variables.scss';
+
+
